flask_app/requirements.txt
# Flask Core
Flask
Flask-SocketIO
Flask-CORS
python-socketio
python-engineio
eventlet

# Existing Dependencies (Reused from Streamlit)
opencv-python
pillow
pyserial
ultralytics
numpy

# Optional but recommended
python-dotenv
